requirements.txt
capstone~=5.0.1
unicorn~=2.0.1.post1
pyelftools~=0.31
keystone~=25.0.0
colorama~=0.4.6
frida~=16.4.8